投稿時ユーザーid登録

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Tweet;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
+use App\Services\TweetService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
@@ -14,9 +16,12 @@ class DeleteController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int) $request->route('tweetId');
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweetId)){
+            throw new AccessDeniedHttpException();
+        }
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
         return redirect()->route('tweet.index')->with('feedback.success', "つぶやきを削除しました");
